refactor out core chart component into core-ui
